Implement task updates with distributed locks
In order to be able to serve the latest output from all worker machines
executing a particular command after they're all completed, we use a
distributed lock.

We keep track of the outputs of each server in turn in a Redis set.
However, before mutating this set, we grab a lock, since we want to
prevent potential spillovers from the next wave of outputs of the
scheduled task (e.g, if a task is scheduled to run every 30 seconds or
less).

Once the size of this set is equal to the number of available servers
that we scheduled the task on, we write to a PostgreSQL table that the
task has been completed, storing in a JSONB column the outputs of all
the workers.